checkpoint: 10-2: hardening logic for failed member replacement

+++ pkg/cmd/staticpodcontroller/staticpodcontroller.go
@@ -12,10 +12,8 @@ import (
 
 	ceoapi "github.com/openshift/cluster-etcd-operator/pkg/operator/api"
 
-	"github.com/openshift/cluster-etcd-operator/pkg/operator/operatorclient"
 	"github.com/openshift/cluster-etcd-operator/pkg/version"
 	"github.com/openshift/library-go/pkg/operator/events"
-	"github.com/openshift/library-go/pkg/operator/v1helpers"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/vincent-petithory/dataurl"
@@ -25,12 +23,11 @@ import (
 	"k8s.io/client-go/util/workqueue"
 	"k8s.io/klog"
 
+	operatorclient "github.com/openshift/client-go/operator/clientset/versioned/typed/operator/v1"
 	mcfgclientset "github.com/openshift/machine-config-operator/pkg/generated/clientset/versioned"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
-	operatorversionedclient "github.com/openshift/client-go/operator/clientset/versioned"
-	operatorv1informers "github.com/openshift/client-go/operator/informers/externalversions"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/informers"
@@ -86,7 +83,7 @@ func (s *podOpts) Run() error {
 	klog.Infof("Version: %+v (%s)", info.GitVersion, info.GitCommit)
 
 	// here
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := context.WithCancel(context.TODO())
 	clientConfig, err := rest.InClusterConfig()
 	if err != nil {
 		return err
@@ -99,30 +96,34 @@ func (s *podOpts) Run() error {
 	if err != nil {
 		return err
 	}
-	operatorConfigClient, err := operatorversionedclient.NewForConfig(clientConfig)
+	operatorClient, err := operatorclient.NewForConfig(clientConfig)
 	if err != nil {
 		return err
 	}
-	operatorConfigInformers := operatorv1informers.NewSharedInformerFactory(operatorConfigClient, 10*time.Minute)
-	operatorClient := &operatorclient.OperatorClient{
-		Informers: operatorConfigInformers,
-		Client:    operatorConfigClient.OperatorV1(),
-	}
+	etcdOperatorClient := operatorClient.Etcds()
 
-	kubeInformerFactory := informers.NewFilteredSharedInformerFactory(clientset, 0, "openshift-etcd", nil)
+	kubeInformerFactory := informers.NewFilteredSharedInformerFactory(clientset, 0, etcdNamespace, nil)
 	nodeName := os.Getenv("NODE_NAME")
 	var localEtcdName string
 	if nodeName != "" {
 		localEtcdName = fmt.Sprintf("etcd-member-%s", nodeName)
 	}
 
+	//TODO: util v6j
+	controllerRef, err := events.GetControllerReferenceForCurrentPod(clientset, etcdNamespace, nil)
+	if err != nil {
+		klog.Warningf("unable to get owner reference (falling back to namespace): %v", err)
+	}
+
+	eventRecorder := events.NewKubeRecorder(clientset.CoreV1().Events(etcdNamespace), "static-pod-controller-"+localEtcdName, controllerRef)
+
 	staticPodController := NewStaticPodController(
-		operatorClient,
+		etcdOperatorClient,
 		kubeInformerFactory,
 		clientset,
 		clientmc,
 		localEtcdName,
-		// ctx.EventRecorder,
+		eventRecorder,
 	)
 
 	kubeInformerFactory.Start(ctx.Done())
@@ -135,7 +136,7 @@ func (s *podOpts) Run() error {
 }
 
 type StaticPodController struct {
-	operatorConfigClient                   v1helpers.OperatorClient
+	etcdOperatorClient                     operatorclient.EtcdInterface
 	podInformer                            corev1informer.SecretInformer
 	kubeInformersForOpenshiftEtcdNamespace cache.SharedIndexInformer
 	clientset                              corev1client.Interface
@@ -148,19 +149,19 @@ type StaticPodController struct {
 }
 
 func NewStaticPodController(
-	operatorConfigClient v1helpers.OperatorClient,
+	etcdOperatorClient operatorclient.EtcdInterface,
 	kubeInformersForOpenshiftEtcdNamespace informers.SharedInformerFactory,
 	clientset corev1client.Interface,
 	clientmc mcfgclientset.Interface,
 	localEtcdName string,
-	// eventRecorder events.Recorder,
+	eventRecorder events.Recorder,
 ) *StaticPodController {
 	c := &StaticPodController{
-		operatorConfigClient: operatorConfigClient,
-		// eventRecorder: eventRecorder.WithComponentSuffix("static-pod-controller"),
-		clientset:     clientset,
-		clientmc:      clientmc,
-		localEtcdName: localEtcdName,
+		etcdOperatorClient: etcdOperatorClient,
+		eventRecorder:      eventRecorder.WithComponentSuffix("static-pod-controller-" + localEtcdName),
+		clientset:          clientset,
+		clientmc:           clientmc,
+		localEtcdName:      localEtcdName,
 
 		queue: workqueue.NewNamedRateLimitingQueue(workqueue.DefaultControllerRateLimiter(), "StaticPodController"),
 	}
@@ -171,42 +172,46 @@ func NewStaticPodController(
 }
 
 func (c *StaticPodController) sync() error {
-	pod, err := c.clientset.CoreV1().Pods("openshift-etcd").Get(c.localEtcdName, metav1.GetOptions{})
+	pod, err := c.clientset.CoreV1().Pods(etcdNamespace).Get(c.localEtcdName, metav1.GetOptions{})
 	if err != nil {
-		klog.Infof("No Pod found in openshift-etcd with name %s", c.localEtcdName)
+		klog.Infof("No Pod found in %s with name %s", etcdNamespace, c.localEtcdName)
 		return err
 	}
 	staticPodPath := fmt.Sprintf("%s/%s", manifestDir, etcdMemberFileName)
 	// If last status is CrashLoopBackOff we perform further inspection to verify.
-	if pod.Status.ContainerStatuses[0].State.Waiting != nil && pod.Status.ContainerStatuses[0].State.Waiting.Reason == "CrashLoopBackOff" {
+	if len(pod.Status.ContainerStatuses) > 0 {
+		if pod.Status.ContainerStatuses[0].State.Waiting != nil && pod.Status.ContainerStatuses[0].State.Waiting.Reason == "CrashLoopBackOff" {
 
-		klog.Warningf("%s is unhealthy", c.localEtcdName)
+			klog.Infof("%s is unhealthy", c.localEtcdName)
 
-		// if the data-dir is missing we are going to give it some time to recover.
-		if _, err := os.Stat(fmt.Sprintf("%s/member/snap", dataDir)); os.IsNotExist(err) {
-			klog.Warningf("data-dir already removed %s", dataDir)
-			return nil
-		}
-
-		if c.IsMemberRemove(c.localEtcdName) {
-			etcdMember, err := c.getMachineConfigData(staticPodPath, "master")
-			if err != nil {
-				return err
-			}
-			klog.Infof("removing static pod %s\n", staticPodPath)
-			if err := os.Remove(staticPodPath); err != nil {
-				return err
+			// if the data-dir is missing we are going to give it some time to recover.
+			if _, err := os.Stat(fmt.Sprintf("%s/member/snap", dataDir)); os.IsNotExist(err) {
+				klog.Warningf("data-dir already removed %s", dataDir)
+				return nil
 			}
 
-			//TODO verify etcd-member container has exited instead of sleep
-			klog.Infof("sleeping for %d\n", 10)
-			time.Sleep(10 * time.Second)
+			if c.IsMemberRemove(c.localEtcdName) {
+				klog.Infof("%s is pending removal", c.localEtcdName)
+				etcdMember, err := c.getMachineConfigData(staticPodPath, "master")
+				if err != nil {
+					klog.Warningf("etcdMember failed %v", err)
+					return err
+				}
+				klog.Infof("removing static pod %s\n", staticPodPath)
+				if err := os.Remove(staticPodPath); err != nil {
+					return err
+				}
 
-			klog.Infof("removing data-dir contents")
-			os.RemoveAll(fmt.Sprintf("%s/member", dataDir))
-			klog.Infof("starting %s", c.localEtcdName)
-			if err := ioutil.WriteFile(staticPodPath, []byte(etcdMember), 0644); err != nil {
-				return err
+				//TODO verify etcd-member container has exited instead of sleep
+				klog.Infof("sleeping for %d\n", 10)
+				time.Sleep(10 * time.Second)
+
+				klog.Infof("removing data-dir contents")
+				os.RemoveAll(fmt.Sprintf("%s/member", dataDir))
+				klog.Infof("starting %s", c.localEtcdName)
+				if err := ioutil.WriteFile(staticPodPath, []byte(etcdMember), 0644); err != nil {
+					return err
+				}
 			}
 		}
 	}
@@ -214,7 +219,10 @@ func (c *StaticPodController) sync() error {
 }
 
 func (c *StaticPodController) IsMemberRemove(name string) bool {
-	members, _ := c.PendingMemberList()
+	members, err := c.PendingMemberList()
+	if err != nil {
+		klog.Errorf("IsMemberRemove: error %v", err)
+	}
 	for _, m := range members {
 		klog.Warningf("IsMemberRemove: checking %v vs %v type = %v\n", m.Name, name, m.Conditions[0].Type)
 		if m.Name == name && m.Conditions[0].Type == ceoapi.MemberRemove {
@@ -226,12 +234,13 @@ func (c *StaticPodController) IsMemberRemove(name string) bool {
 
 func (c *StaticPodController) PendingMemberList() ([]ceoapi.Member, error) {
 	configPath := []string{"cluster", "pending"}
-	operatorSpec, _, _, err := c.operatorConfigClient.GetOperatorState()
+	operatorSpec, err := c.etcdOperatorClient.Get("cluster", metav1.GetOptions{})
 	if err != nil {
 		return nil, err
 	}
+
 	config := map[string]interface{}{}
-	if err := json.NewDecoder(bytes.NewBuffer(operatorSpec.ObservedConfig.Raw)).Decode(&config); err != nil {
+	if err := json.NewDecoder(bytes.NewBuffer(operatorSpec.Spec.ObservedConfig.Raw)).Decode(&config); err != nil {
 		klog.V(4).Infof("decode of existing config failed with error: %v", err)
 	}
 	data, exists, err := unstructured.NestedSlice(config, configPath...)
